Server/requirements.txt
google-generativeai
python-dotenv
watchdog